pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-ai-gen-ppt"
version = "1.0.0"
description = "基于AI接口的PDF内容处理与PPT自动生成系统"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "PDF-AI-GEN_PPT Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "pdfminer.six>=20221105",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    "python-pptx>=0.6.21",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "chardet>=5.2.0",
]

[project.scripts]
pdf-ai-gen = "src.main:main"

[project.urls]
Homepage = "https://github.com/your-repo/pdf-ai-gen-ppt"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
